requirements.txt
asgiref==3.2.7
atomicwrites==1.4.0
attrs==19.3.0
colorama==0.4.3
Django==3.0.6
django-redis==4.11.0
djangorestframework==3.11.0
importlib-metadata==1.6.0
jsonschema==3.2.0
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytz==2020.1
redis==3.5.0
six==1.14.0
sortedcontainers==2.1.0
sqlparse==0.3.1
wcwidth==0.1.9
zipp==3.1.0
